project restructure!
notes:
- semver wasn't increased cuz im not sure if it should be yet
- api compatability with old code was tested, tests are also all successful
- formatted with the provided rustfmt config
- as outlined in the edits, some fields were made public because i \
    couldn't figure out how to make them work without doing so, they \
    aren't visible in the docs but i'm still not sure.

edits:
- it will now compile on stable rust
- in order to make it compile on stable rust a new function that mimicked \
    the assume_init function had to be made "assume_init_box"
- all the code was separated into 3 modules, common, encoder and decoder \
    (visibilities and api were kept as is, old source code should still \
    be compatible with the restructure)
- function "new" added for `OggPacket` to make stuff cleaner
- `VorbisInfo`'s (in common/vorbis_data.rs:20-21) field was made public
- `VorbisComments`'s (in common/vorbis_data.rs:196-197) field was made public
- instead of collect_into, extend was used (decoder/decoder_data.rs:39-41)
- `VorbisEncodingState`'s (encoder/encoder_data.rs:13-18) fields were made public
- all occurences of .assume_init() being used on Box objects were replaced \
    with calls to the new "assume_init_box" function
