+++ packages/vorbis/src/decoder.rs
@@ -1,0 +1,5 @@
+mod decoder_data;
+pub use decoder_data::VorbisAudioSamples;
+
+mod decoder;
+pub use decoder::VorbisDecoder;
