+++ packages/vorbis/src/encoder.rs
@@ -1,0 +1,5 @@
+mod encoder_data;
+pub use encoder_data::{VorbisBitrateManagementStrategy, VorbisEncodingState};
+
+mod encoder;
+pub use encoder::VorbisEncoder;
